+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* `local_mock()` and `with_mock()` have been deprecated in favour of
+  `local_mocked_responses()` and `with_mocked_responses()` (#301).
+
 * New `oauth_cache_path()` returns the path that httr2 uses for caching OAuth
   tokens. Additionally, you can now change the cache location by setting the
   `HTTR2_OAUTH_CACHE` env var.